move enum to devices and rename
