+++ examples/job.py
@@ -13,8 +13,9 @@
 
 import os
 
-from braket.aws import AwsDevice, AwsQuantumJob, BraketDevices
+from braket.aws import AwsDevice, AwsQuantumJob
 from braket.circuits import Circuit
+from braket.devices import Devices
 from braket.jobs import save_job_result
 
 
@@ -36,7 +37,7 @@ def run_job():
 
 if __name__ == "__main__":
     job = AwsQuantumJob.create(
-        device=BraketDevices.Amazon.SV1,
+        device=Devices.Amazon.SV1,
         source_module="job.py",
         entry_point="job:run_job",
         wait_until_complete=True,
